requirements.txt
pillow
escpos
pyusb
pypdf
cython